+++ jflex/src/main/java/jflex/ErrorMessages.java
@@ -141,8 +141,6 @@ public class ErrorMessages {
   public static ErrorMessages DOT_BAR_NEWLINE_DOES_NOT_MATCH_ALL_CHARS = new ErrorMessages("DOT_BAR_NEWLINE_DOES_NOT_MATCH_ALL_CHARS");
   public static ErrorMessages PROPS_ARG_REQUIRES_UNICODE_VERSION = new ErrorMessages("PROPS_ARG_REQUIRES_UNICODE_VERSION");
   public static ErrorMessages EMITTING_INPUTSTREAM_CTOR = new ErrorMessages("EMITTING_INPUTSTREAM_CTOR");
-  public static ErrorMessages SWITCH_METHOD_DEPRECATED = new ErrorMessages("SWITCH_METHOD_DEPRECATED");
-  public static ErrorMessages TABLE_METHOD_DEPRECATED = new ErrorMessages("TABLE_METHOD_DEPRECATED");
   public static ErrorMessages IMPOSSIBLE_CHARCLASS_RANGE = new ErrorMessages("IMPOSSIBLE_CHARCLASS_RANGE");
   public static ErrorMessages CODEPOINT_OUT_OF_RANGE = new ErrorMessages("CODEPOINT_OUT_OF_RANGE");
 }